+++ assets/jsx/workflow-editor/components/data-fields/post-data/post-author.jsx
@@ -38,7 +38,7 @@ const getAuthors = () => {
 const getAuthorOptions = (authors) => {
     return authors.map(author => ({
         value: author.id,
-        label: author.name + ' (' + author.email + ')',
+        label: author.name,
     }));
 };